Made adjustments to code based on code review
Made adjustments to code based on code review. Moved file name changes to the beginning of each cell where they are needed and saved the file names as a variable to be joined to the file path deeper in the code. Made corresponding changes to the annotations of the various cells where the code adjustments had been made.