Made changes to mallet model run time improvements
When saving the mallet model it was not saving the temp files so that you could just run the results notebook on what was created by the data clean notebook after you logged out of ReD. Added the prefix parameter so it would save the temp files. This caused parts of the results notebook to run much slower, so I also made changes to improve run time.