Corrected potential issue
Corrected issue where the code might error if one of the data frames does not have all the topics listed which is possible if one of the topics is not dominant in any of the documents. Changed the code to use the length of the show_topics function instead of the data frame.